Update Get-GPPPassword.ps1
The original script couldn't deal with multiple objects contained within the XML, e.g., one policy setting passwords for 2 user accounts. I've updated the script so that it handles these situations properly. Also made a few other minor changes: hashtable order, and general formatting.

I've never live edited a script prior to this and hadn't planned on submitting it like this. The changes in formatting we're really only my preference to make it easier for me to read. 

My apologies if that's frowned upon as I never intended to make this public to anyone other than the author.