Add additional search entry for 'Get-UnattendedInstallFile'
Based on  https://docs.microsoft.com/en-us/windows-hardware/manufacture/desktop/windows-setup-automation-overview, under the Implicit Answer File Search Order section, an additional location for an answer file may exist at HKLM\SYSTEM\Setup\UnattendFile. As pointed in the article, the full path to the file is stored under the aforementioned registry key. The file does not need to be named as Unattend.xml or the like, and can be located anywhere on the drive.