Allow DLL loading without calling any function.
There are many DLL payloads that all start up from DllMain(DLL_PROCESS_ATTACH). They do not export any method so there is no need to call one. In order to work with those payloads, the script had to be modified so that `-FuncReturnType` could accept value of `None` and therefore avoid calling any function in the DLL. Thus of course, leaving all the job for the DLL's entrypoint routine.